saving place, mostly builds, needs to actually start boosting the limit based on fallback limit

+++ service/matching/tasklist/task_list_manager.go
@@ -471,7 +471,7 @@ func (c *taskListManagerImpl) DescribeTaskList(includeTaskListStatus bool) *type
 		ReadLevel:        c.taskAckManager.GetReadLevel(),
 		AckLevel:         c.taskAckManager.GetAckLevel(),
 		BacklogCountHint: backlogCount,
-		RatePerSecond:    c.matcher.Rate(),
+		RatePerSecond:    float64(c.matcher.Rate()),
 		TaskIDBlock: &types.TaskIDBlock{
 			StartID: taskIDBlock.start,
 			EndID:   taskIDBlock.end,
